message/kafka_module/pyproject.toml
[tool.poetry]
name = "kafka-module"
version = "0.1.2"
description = ""
authors = ["parkyoungjin-888 <pyj8813@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
pydantic = "2.7.4"
kafka-python-ng = "2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
